apply authorization for task APIs

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { Repository } from 'typeorm';
 // import { Task } from './task.model';
 import { v4 } from 'uuid';
@@ -16,31 +17,31 @@ export class TasksService {
         private taskRepository: TaskRepository
     ){}
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    getTasks = (filterTaskDto: FilterTaskDto): Promise<Task[]> => {
-        return this.taskRepository.getTasks(filterTaskDto);
+    getTasks = (filterTaskDto: FilterTaskDto, user: User): Promise<Task[]> => {
+        return this.taskRepository.getTasks(filterTaskDto, user);
     }
-    
-    async getTaskById(id: number): Promise<Task> {
-        const result = await this.taskRepository.findOne(id);
+
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const result = await this.taskRepository.findOne({ where: { id: id, userId: user.id } });
         if(!result){
             throw new NotFoundException(`Task with id ${id} not found`);
         }
         return result;
     }
 
-    async updateTaskById(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskById(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         const result = await this.taskRepository.save(task); //or task.save()
         return result;
     }
 
-    async deleteTaskById(id: number): Promise<void>{
-        const result = await this.taskRepository.delete(id);
+    async deleteTaskById(id: number, user: User): Promise<void>{
+        const result = await this.taskRepository.delete({ id, userId: user.id });
         if(result.affected === 0){
             throw new NotFoundException(`Task with ID "${id}" not found`);
         }
